+++ src/widgets/layers/PNotifyAsset.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace lo\modules\noty\widgets\layers;
+
+use yii\web\AssetBundle;
+
+/**
+ * Class PNotifyAsset
+ * @package lo\modules\noty\widgets\layers
+ */
+class PNotifyAsset extends AssetBundle
+{
+    /** @var string  */
+    public $sourcePath = '@bower/pnotify/dist';
+
+    /** @var array $css */
+    public $css = [
+        'pnotify.css',
+        'pnotify.brighttheme.css',
+        'pnotify.buttons.css',
+        'pnotify.history.css',
+        'pnotify.mobile.css',
+    ];
+
+    /** @var array $js */
+    public $js = [
+        'pnotify.js',
+        'pnotify.animate.js',
+        'pnotify.buttons.js',
+        'pnotify.desktop.js',
+        'pnotify.history.js',
+        'pnotify.mobile.js',
+        'pnotify.nonblock.js',
+    ];
+
+    /** @var array $depends */
+    public $depends = [
+        'yii\web\JqueryAsset'
+    ];
+}