+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+* 2.3 [30 March 2016]
+    - `enh` Added new layer - PNotify
+    - `enh` Added new layer - Jquery Notify
+    - `fix` Fixed translations
+
 * 2.2 [29 March 2016]
     - `enh` Added new layer - Noty.js
 
@@ -9,7 +14,6 @@
 * 2.0 [26 March 2016]
     - `enh` Added new entity "Layers".
     - `enh` Delete widgets from composer.
-    - `fix` Fixed translations sourceLanguage (thezilla-)
 
 * 1.1 [22 March 2016]
     - `enh` Rename Noty to Wrapper. First stable release